+++ CHANGELOG.md
@@ -6,7 +6,9 @@
 * Feature: Check for unknown option and free space errors
 * Feature: Comment lines and empty lines allowed in profiles
 * Feature: Pushover support in .rsink/tools
-* Feature: Options system (h = help, p = pushover)
+* Feature: Options system
+* Feature: Pushover (`-p`)
+* Feature: Help (`-h`) and version (`-v`) options
 
 # v0.1 (December 12, 2013)
 * Initial release